refactor: organize src/game/ into puzzles/<family>/ and traps/ subfolders
Sumplete, Crocodile, and Tableau each get their own folder under
src/game/puzzles/ holding generation + DDD state + checks together,
with shared plugin/registry infra staying at src/game/puzzles/ root.
Trap plugin/registry/config move to src/game/traps/ the same way.
Journey/level/site-assembly files stay flat at src/game/ since they
aren't puzzle/trap-type specific.
