+++ src/app/PuzzleFamilies/Sumplete/plugin.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable react-refresh/only-export-components -- side-effect registration file */
 import { type FC } from "react"
-import { registerPuzzle } from "@/game/puzzleRegistry"
-import { generateSumplete, type SumpleteGrid } from "@/game/generateSumplete"
+import { registerPuzzle } from "@/game/puzzles/puzzleRegistry"
+import { generateSumplete, type SumpleteGrid } from "@/game/puzzles/sumplete/generateSumplete"
 import { SumpletePuzzle } from "./SumpletePuzzle"
-import type { PuzzleSettings } from "@/game/puzzlePlugin"
+import type { PuzzleSettings } from "@/game/puzzles/puzzlePlugin"
 
 const SumpleteComponent: FC<{ puzzle: SumpleteGrid; settings: PuzzleSettings; onSolved: () => void }> = ({
   puzzle,
