+++ src/app/PyramidLevel/inventoryLootLogic.ts
@@ -1,7 +1,10 @@
 import { type CombinedJourneyState } from "@/app/state/useJourneys"
 import { journeys, type TreasureTombJourney } from "@/data/journeys"
 import { tableauLevels } from "@/data/tableaus"
-import { buildTombCalculationSettings, generateRewardCalculation } from "@/game/generateRewardCalculation"
+import {
+  buildTombCalculationSettings,
+  generateRewardCalculation,
+} from "@/game/puzzles/tableau/generateRewardCalculation"
 import { generateNewSeed, mulberry32, shuffle } from "@/game/random"
 import { getItemFirstLevel } from "@/data/itemLevelLookup"
 import { type Difficulty, difficultyCompare } from "@/data/difficultyLevels"
