+++ src/app/pages/TableauInventory.tsx
@@ -3,7 +3,10 @@ import { useJourneys, type CombinedJourneyState } from "../state/useJourneys"
 import { journeys, type TreasureTombJourney } from "@/data/journeys"
 import { useTableauTranslations } from "@/data/useTableauTranslations"
 import { generateNewSeed, mulberry32 } from "@/game/random"
-import { buildTombCalculationSettings, generateRewardCalculation } from "@/game/generateRewardCalculation"
+import {
+  buildTombCalculationSettings,
+  generateRewardCalculation,
+} from "@/game/puzzles/tableau/generateRewardCalculation"
 import { useInventory } from "../Inventory/useInventory"
 import { getInventoryItemById } from "@/data/inventory"
 import { getItemFirstLevel } from "@/data/itemLevelLookup"
