+++ src/game/puzzles/crocodile/crocodileState.spec.ts
