+++ src/game/puzzles/crocodile/crocodileState.ts
