+++ src/game/puzzles/puzzlePlugin.ts
