+++ src/game/puzzles/puzzleRegistry.ts
