+++ src/game/puzzles/sumplete/sumpleteStatus.ts
