+++ src/game/puzzles/tableau/generateRewardCalculation.ts
@@ -5,7 +5,7 @@ import {
   type Formula,
   type FormulaSettings,
   type Operation,
-} from "../app/Formulas/formulas"
+} from "@/app/Formulas/formulas"
 
 /**
  * Describe a reward calculation.
