+++ src/game/puzzles/tableau/tableauPuzzleState.spec.ts
