+++ src/game/puzzles/tableau/tableauPuzzleState.ts
