+++ src/game/traps/trapConfig.ts
