+++ src/game/traps/trapPlugin.ts
