+++ src/game/traps/trapRegistry.ts
