+++ src/ui/organisms/TombPuzzleView.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite"
 import { generateNewSeed, mulberry32 } from "@/game/random"
 import { journeys } from "@/data/journeys"
 import { hashString } from "@/support/hashString"
-import { generateRewardCalculation } from "@/game/generateRewardCalculation"
+import { generateRewardCalculation } from "@/game/puzzles/tableau/generateRewardCalculation"
 import { useTableauTranslations } from "@/data/useTableauTranslations"
 import { resolveHieroglyphSymbol } from "@/data/resolveHieroglyphSymbol"
 import { TombPuzzleView } from "./TombPuzzleView"
