+++ src/ui/organisms/TombTableau.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite"
 import { generateNewSeed, mulberry32 } from "@/game/random"
 import { journeys, type TreasureTombJourney } from "@/data/journeys"
 import { hashString } from "@/support/hashString"
-import { generateRewardCalculation } from "@/game/generateRewardCalculation"
+import { generateRewardCalculation } from "@/game/puzzles/tableau/generateRewardCalculation"
 import { useMemo } from "react"
 import { TombTableau } from "./TombTableau"
 import { createPositionOverview } from "@/app/Formulas/filledPositions"
